+++ 0x30/Menu.h
@@ -16,7 +16,7 @@ public:
     void Start(double t);
     Screens Update(double t, double dt);
     void Draw(double t);
-    int SelectedLevel() const { return currentSelection_ + 1; }
+    size_t SelectedLevel() const { return currentSelection_ + 1; }
     Mode SelectedMode() const { return mode_; }
 
 private:
@@ -27,8 +27,8 @@ private:
     TextRenderer textRenderer_;
 
     double screenStartTime_{0};
-    int currentSelection_{0};
-    int firstVisibleLevel_{0};
+    size_t currentSelection_{0};
+    size_t firstVisibleLevel_{0};
     Mode mode_{Mode::ENDLESS};
 
     static constexpr int visibleLevels_{10};
